updates with api call, lower bound, output
I’ve made some changes: stored value in a flask session so we don’t get API failures (but I can also implement a wait function if it still crashes), made the hossain page output parameters and measurement in foco cups <- but this function is not in place yet, I put in a random placeholder function bc i didnt know how many grams can fit into a foco cup, set lower bound to the hossain function
