--- ENGS90/hello.py
@@ -1,7 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-@app.route("/")
-def hello_world():
-    return "Hello, World!"
